fix: move shared messages
